oauth2/srcs/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==11.0.0
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.1
pytz==2024.1
qrcode==8.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
